chore: convert apply/mark files
