--- src/apply/mark/index.js
@@ -1,9 +1,0 @@
-const addMark = require('./addMark');
-const removeMark = require('./removeMark');
-
-const mappers = {
-    add_mark: addMark,
-    remove_mark: removeMark,
-};
-
-module.exports = mappers;
