+++ src/apply/mark/index.ts
@@ -1,0 +1,7 @@
+import addMark from './addMark';
+import removeMark from './removeMark';
+
+export default {
+    add_mark: addMark,
+    remove_mark: removeMark,
+};
